Implement temperature functionality as a plugin
